Handle omit special case for OFFSET statement
Because constraints are considered as a hint, we should not omit the
constraints verification by SQLite. Therefore, the omit flag must be
set to 0.
However, if an OFFSET statement is used, the omit flag must be set to 1
because the rows can't be offsetted twice (or some rows will be missing)