Fix SQL joins in virtual table
BestIndex method of a virtual table in SQLite is called several times
to find the best querying strategy.
To note that a strategy is wrong, it should return an error of
SQLITE_CONSTRAINT.

The current implementation of the bestIndex method in the virtual table
fails at the first error, and does not try the other strategies.

Therefore, SQL joins are not correctly handled in the virtual table.